+++ common/src/main/java/com/evandev/remi/mixin/emi/EmiScreenManagerAccessor.java
@@ -1,0 +1,16 @@
+package com.evandev.remi.mixin.emi;
+
+import dev.emi.emi.screen.EmiScreenManager;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.List;
+
+@Mixin(value = EmiScreenManager.class, remap = false)
+public interface EmiScreenManagerAccessor {
+
+    @Accessor("panels")
+    static List<EmiScreenManager.SidebarPanel> getPanels() {
+        throw new UnsupportedOperationException();
+    }
+}