+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.3.0] - 2026-08-12
+
+### Added
+
+- Added support for additional workstations in the Workstation tab.
+- Added support for stack groups and creative tab filtering in other tabs.
+- Added additional config options for sidebar customization.
+
 ## [4.2.1] - 2026-08-11
 
 ### Added